expose anthropic channel end-to-end and fix billing/concurrency/sql
* unify BreakerSnapshot across channels so admin/ws share one wire type
* admin/upstream breaker_get and breaker_post cover anthropic with reset/disable
* dashboard channel-aware modals, per-channel test model, hide allow_fast for anthropic
* anthropic key_pool exposes snapshot_breakers / reset_breaker / force_disable_breaker
* fix ANTHROPIC_OPUS pricing to current rate
* concurrency GC keeps entries with count > 0 (active long-running requests)
* SQL totals query coalesces channel_kind across errors-only channels
* simplify exa key picker to take a slice
* migrate api_keys/usage_logs/error_logs/upstream_keys id to BIGINT
* drop unused tls_* config fields
* anthropic billing now logs the client request body
* shutdown drains in-flight connections via inflight tracker instead of fixed sleep

+++ src/db/errors.rs
@@ -27,8 +27,8 @@ pub async fn insert_error(db: &Db, input: ErrorLogInput) -> AppResult<i64> {
     .fetch_one(db.pool())
     .await?;
 
-    let id: i32 = row.try_get("id")?;
-    Ok(id as i64)
+    let id: i64 = row.try_get("id")?;
+    Ok(id)
 }
 
 pub async fn cleanup_old_errors(db: &Db, key_name: &str) -> AppResult<()> {
@@ -103,7 +103,7 @@ pub async fn get_error_by_id(db: &Db, id: i64) -> AppResult<Option<ErrorLog>> {
         "SELECT id, time, key_name, status, path, model, request_body, response_body, ip, channel_kind \
          FROM error_logs WHERE id = $1",
     )
-    .bind(id as i32)
+    .bind(id)
     .fetch_optional(db.pool())
     .await?;
     Ok(row)
